boomertv2/requirements.txt
psycopg2-binary==2.9.7
PyYAML==6.0.1
neo4j==5.15.0
sentence-transformers==2.2.2
openai==1.3.0
numpy==1.24.3
mcp>=0.1.0